pyproject.toml
[tool.poetry]
name = "stock-analysis-app"
version = "0.1.0"
description = ""
authors = ["Vimal Kansal <vimalkansal@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.39.0"
yfinance = "^0.2.44"
pandas = "^2.2.3"
numpy = "^1.24.0"
plotly = "^5.24.1"
ta = "^0.11.0"
textblob = "^0.18.0.post0"
requests = "^2.32.3"
openai = "^1.51.0"
langchain = "^0.3.2"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.1"
langchain-openai = "^0.2.1"
aiofiles = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
